+++ ExpandWorldPrefabs/DelayedSpawn.cs
@@ -1,20 +1,20 @@
 using System.Collections.Generic;
-using Service;
+using Data;
 using UnityEngine;
 
 namespace ExpandWorld.Prefab;
 
-public class DelayedSpawn(float delay, Vector3 pos, Quaternion rot, int prefab, long originalOwner, ZDOData? data)
+public class DelayedSpawn(float delay, Vector3 pos, Quaternion rot, int prefab, long originalOwner, DataEntry? data, Dictionary<string, string> parameters, bool triggerRules)
 {
   private static readonly List<DelayedSpawn> Spawns = [];
-  public static void Add(float delay, Vector3 pos, Quaternion rot, int prefab, long originalOwner, ZDOData? data)
+  public static void Add(float delay, Vector3 pos, Quaternion rot, int prefab, long originalOwner, DataEntry? data, Dictionary<string, string> parameters, bool triggerRules)
   {
     if (delay <= 0f)
     {
-      Manager.CreateObject(prefab, pos, rot, originalOwner, data);
+      Manager.CreateObject(prefab, pos, rot, originalOwner, data, parameters, triggerRules);
       return;
     }
-    Spawns.Add(new(delay, pos, rot, prefab, originalOwner, data));
+    Spawns.Add(new(delay, pos, rot, prefab, originalOwner, data, parameters, triggerRules));
   }
   public static void Execute(float dt)
   {
@@ -35,11 +35,13 @@ public class DelayedSpawn(float delay, Vector3 pos, Quaternion rot, int prefab, 
   private readonly Quaternion Rot = rot;
   private readonly int Prefab = prefab;
   private readonly long OriginalOwner = originalOwner;
-  private readonly ZDOData? Data = data;
+  private readonly DataEntry? Data = data;
   public float Delay = delay;
+  public Dictionary<string, string> Parameters = parameters;
+  public bool TriggerRules = triggerRules;
 
   public void Execute()
   {
-    Manager.CreateObject(Prefab, Pos, Rot, OriginalOwner, Data);
+    Manager.CreateObject(Prefab, Pos, Rot, OriginalOwner, Data, Parameters, TriggerRules);
   }
 }