+++ ExpandWorldPrefabs/Helper.cs
@@ -1,20 +1,29 @@
 using System;
+using System.Collections.Generic;
 
 namespace ExpandWorld.Prefab;
 
 public class Helper2
 {
 
-  public static string ReplaceParameters(string str, string prefab, string parameter)
+  public static string ReplaceParameters(string str, Dictionary<string, string> parameters)
   {
-    var split = parameter.Split(' ');
-    return str.Replace("<prefab>", prefab)
-      .Replace("<par0>", split.Length > 0 ? split[0] : "")
-      .Replace("<par1>", split.Length > 1 ? split[1] : "")
-      .Replace("<par2>", split.Length > 2 ? split[2] : "")
-      .Replace("<par3>", split.Length > 3 ? split[3] : "")
-      .Replace("<par4>", split.Length > 4 ? split[4] : "")
-      .Replace("<par>", parameter);
+    foreach (var pair in parameters)
+      str = str.Replace(pair.Key, pair.Value);
+    return str;
+  }
+  public static Dictionary<string, string> CreateParameters(string prefab, string args)
+  {
+    var split = args.Split(' ');
+    return new Dictionary<string, string> {
+      { "<prefab>", prefab },
+      { "<par0>", split.Length > 0 ? split[0] : "" },
+      { "<par1>", split.Length > 1 ? split[1] : "" },
+      { "<par2>", split.Length > 2 ? split[2] : "" },
+      { "<par3>", split.Length > 3 ? split[3] : "" },
+      { "<par4>", split.Length > 4 ? split[4] : "" },
+      { "<par>", args }
+    };
   }
   public static bool CheckWild(string wild, string str)
   {