+++ ExpandWorldPrefabs/ObjectsFiltering.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
-using ExpandWorldData;
+using Service;
 using UnityEngine;
 
 namespace ExpandWorld.Prefab;
@@ -8,51 +8,51 @@ namespace ExpandWorld.Prefab;
 
 public class ObjectsFiltering
 {
-  public static ZDO[] GetNearby(int limit, Object[] objects, ZDO zdo, string name, string parameter)
+  public static ZDO[] GetNearby(int limit, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
   {
     if (objects.Length == 0) return [];
     var maxRadius = objects.Max(o => o.MaxDistance);
     var indices = GetSectorIndices(zdo.m_position, maxRadius);
-    return GetObjects(limit, indices, objects, zdo, name, parameter);
+    return GetObjects(limit, indices, objects, zdo, parameters);
   }
-  private static ZDO[] GetObjects(int limit, HashSet<int> indices, Object[] objects, ZDO zdo, string name, string parameter)
+  private static ZDO[] GetObjects(int limit, HashSet<int> indices, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
   {
     var pos = zdo.m_position;
     var zm = ZDOMan.instance;
-    var query = indices.SelectMany(z => zm.m_objectsBySector[z]).Where(z => z != zdo && objects.Any(o => o.IsValid(z, pos, name, parameter)));
+    var query = indices.SelectMany(z => zm.m_objectsBySector[z]).Where(z => z != zdo && objects.Any(o => o.IsValid(z, pos, parameters)));
     if (limit > 0)
       query = query.OrderBy(z => Utils.DistanceXZ(z.m_position, pos)).Take(limit);
     return query.ToArray();
   }
 
-  public static bool HasNearby(Range<int>? limit, Object[] objects, ZDO zdo, string name, string parameter)
+  public static bool HasNearby(Range<int>? limit, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
   {
     if (objects.Length == 0) return true;
     var maxRadius = objects.Max(o => o.MaxDistance);
     var indices = GetSectorIndices(zdo.m_position, maxRadius);
     if (limit == null)
-      return HasAllObjects(indices, objects, zdo, name, parameter);
+      return HasAllObjects(indices, objects, zdo, parameters);
     else
-      return HasLimitObjects(indices, limit, objects, zdo, name, parameter);
+      return HasLimitObjects(indices, limit, objects, zdo, parameters);
   }
-  public static bool HasNotNearby(Range<int>? limit, Object[] objects, ZDO zdo, string name, string parameter)
+  public static bool HasNotNearby(Range<int>? limit, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
   {
     if (objects.Length == 0) return true;
     var maxRadius = objects.Max(o => o.MaxDistance);
     var indices = GetSectorIndices(zdo.m_position, maxRadius);
     if (limit == null)
-      return !HasAllObjects(indices, objects, zdo, name, parameter);
+      return !HasAllObjects(indices, objects, zdo, parameters);
     else
-      return !HasLimitObjects(indices, limit, objects, zdo, name, parameter);
+      return !HasLimitObjects(indices, limit, objects, zdo, parameters);
   }
 
-  private static bool HasAllObjects(HashSet<int> indices, Object[] objects, ZDO zdo, string name, string parameter)
+  private static bool HasAllObjects(HashSet<int> indices, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
   {
     var pos = zdo.m_position;
     var zm = ZDOMan.instance;
-    return objects.All(o => indices.Any(z => zm.m_objectsBySector[z].Any(z => o.IsValid(z, pos, name, parameter) && z != zdo)));
+    return objects.All(o => indices.Any(z => zm.m_objectsBySector[z].Any(z => o.IsValid(z, pos, parameters) && z != zdo)));
   }
-  private static bool HasLimitObjects(HashSet<int> indices, Range<int> limit, Object[] objects, ZDO zdo, string name, string parameter)
+  private static bool HasLimitObjects(HashSet<int> indices, Range<int> limit, Object[] objects, ZDO zdo, Dictionary<string, string> parameters)
   {
     var pos = zdo.m_position;
     var counter = 0;
@@ -61,7 +61,7 @@ public class ObjectsFiltering
     {
       foreach (var z in ZDOMan.instance.m_objectsBySector[i])
       {
-        var valid = objects.FirstOrDefault(o => o.IsValid(z, pos, name, parameter) && z != zdo);
+        var valid = objects.FirstOrDefault(o => o.IsValid(z, pos, parameters) && z != zdo);
         if (valid == null) continue;
         counter += valid.Weight;
         if (useMax && limit.Max < counter) return false;