+++ ExpandWorldPrefabs/PrefabManager.cs
@@ -1,4 +1,6 @@
+using System.Collections.Generic;
 using System.Linq;
+using Data;
 using Service;
 using UnityEngine;
 
@@ -7,58 +9,59 @@ namespace ExpandWorld.Prefab;
 public class Manager
 {
 
-  public static void Handle(ActionType type, string parameter, ZDO zdo, ZDO? source = null)
+  public static void Handle(ActionType type, string args, ZDO zdo, ZDO? source = null)
   {
     // Already destroyed before.
     if (ZDOMan.instance.m_deadZDOs.ContainsKey(zdo.m_uid)) return;
     if (!ZNet.instance.IsServer()) return;
     var name = ZNetScene.instance.GetPrefab(zdo.m_prefab)?.name ?? "";
-    var info = InfoSelector.Select(type, zdo, name, parameter, source);
+    var parameters = Helper2.CreateParameters(name, args);
+    var info = InfoSelector.Select(type, zdo, args, parameters, source);
     if (info == null) return;
-    Commands.Run(info, zdo, name, parameter, source);
-    HandleSpawns(info, zdo, name, parameter);
-    Poke(info, zdo, name, parameter);
+    Commands.Run(info, zdo, parameters, source);
+    HandleSpawns(info, zdo, parameters);
+    Poke(info, zdo, parameters);
     if (info.Drops)
       SpawnDrops(zdo);
     // Original object was regenerated to apply data.
     if (info.Remove || info.Data != "")
-      RemoveZDO(zdo);
+      DelayedRemove.Add(info.RemoveDelay, zdo);
   }
-  private static void HandleSpawns(Info info, ZDO zdo, string name, string parameter)
+  private static void HandleSpawns(Info info, ZDO zdo, Dictionary<string, string> parameters)
   {
     // Original object must be regenerated to apply data.
     var regenerateOriginal = !info.Remove && info.Data != "";
     if (info.Spawns.Length == 0 && info.Swaps.Length == 0 && !regenerateOriginal) return;
 
-    var customData = ZDOData.Create(info.Data);
+    var customData = DataHelper.Get(info.Data);
     foreach (var p in info.Spawns)
-      CreateObject(p, name, parameter, zdo, customData);
+      CreateObject(p, zdo, customData, parameters, info.TriggerRules);
 
     if (info.Swaps.Length == 0 && !regenerateOriginal) return;
-    var data = ZDOData.Merge(new("", zdo), customData);
+    var data = DataHelper.Merge(new DataEntry(zdo), customData);
 
     foreach (var p in info.Swaps)
-      CreateObject(p, name, parameter, zdo, data);
+      CreateObject(p, zdo, data, parameters, info.TriggerRules);
     if (regenerateOriginal)
-      CreateObject(zdo, data);
+      CreateObject(zdo, data, parameters, false);
   }
-  private static void RemoveZDO(ZDO zdo)
+  public static void RemoveZDO(ZDO zdo)
   {
     ZDOMan.instance.m_deadZDOs[zdo.m_uid] = ZNet.instance.GetTime().Ticks;
     zdo.SetOwner(ZDOMan.instance.m_sessionID);
     ZDOMan.instance.DestroyZDO(zdo);
   }
-  public static void CreateObject(Spawn spawn, string name, string parameter, ZDO originalZdo, ZDOData? data)
+  public static void CreateObject(Spawn spawn, ZDO originalZdo, DataEntry? data, Dictionary<string, string> parameters, bool triggerRules)
   {
     var pos = originalZdo.m_position;
     var rot = originalZdo.GetRotation();
     pos += rot * spawn.Pos;
     rot *= spawn.Rot;
-    data = ZDOData.Merge(data, ZDOData.Create(spawn.Data));
-    DelayedSpawn.Add(spawn.Delay, pos, rot, spawn.GetPrefab(name, parameter), originalZdo.GetOwner(), data);
+    data = DataHelper.Merge(data, DataHelper.Get(spawn.Data));
+    DelayedSpawn.Add(spawn.Delay, pos, rot, spawn.GetPrefab(parameters), originalZdo.GetOwner(), data, parameters, triggerRules);
   }
-  public static void CreateObject(ZDO originalZdo, ZDOData? data) => CreateObject(originalZdo.m_prefab, originalZdo.m_position, originalZdo.GetRotation(), originalZdo.GetOwner(), data);
-  public static void CreateObject(int prefab, Vector3 pos, Quaternion rot, long owner, ZDOData? data)
+  public static void CreateObject(ZDO originalZdo, DataEntry? data, Dictionary<string, string> parameters, bool triggerRules) => CreateObject(originalZdo.m_prefab, originalZdo.m_position, originalZdo.GetRotation(), originalZdo.GetOwner(), data, parameters, triggerRules);
+  public static void CreateObject(int prefab, Vector3 pos, Quaternion rot, long owner, DataEntry? data, Dictionary<string, string> parameters, bool triggerRules)
   {
     if (prefab == 0) return;
     var obj = ZNetScene.instance.GetPrefab(prefab);
@@ -67,8 +70,8 @@ public class Manager
       EWP.LogError($"Can't spawn missing prefab: {prefab}");
       return;
     }
-    // Prefab hash must be zero to avoid triggering HandleCreated.
-    var zdo = ZDOMan.instance.CreateNewZDO(pos, 0);
+    // Prefab hash is used to check whether to trigger rules.
+    var zdo = ZDOMan.instance.CreateNewZDO(pos, triggerRules ? prefab : 0);
     zdo.Persistent = view.m_persistent;
     zdo.Type = view.m_type;
     zdo.Distant = view.m_distant;
@@ -86,7 +89,7 @@ public class Manager
       owner = closestClient?.m_peer.m_uid ?? 0;
     }
     zdo.SetOwnerInternal(owner);
-    data?.Write(zdo);
+    data?.Write(parameters ?? [], zdo);
   }
 
   public static void SpawnDrops(ZDO zdo)
@@ -109,10 +112,10 @@ public class Manager
     UnityEngine.Object.Destroy(obj);
   }
 
-  public static void Poke(Info info, ZDO zdo, string name, string parameter)
+  public static void Poke(Info info, ZDO zdo, Dictionary<string, string> parameters)
   {
-    var zdos = ObjectsFiltering.GetNearby(info.PokeLimit, info.Pokes, zdo, name, parameter);
-    var pokeParameter = Helper2.ReplaceParameters(info.PokeParameter, name, parameter);
+    var zdos = ObjectsFiltering.GetNearby(info.PokeLimit, info.Pokes, zdo, parameters);
+    var pokeParameter = Helper2.ReplaceParameters(info.PokeParameter, parameters);
     foreach (var z in zdos)
       Handle(ActionType.Poke, pokeParameter, z);
   }