+++ publish/CHANGELOG.md
@@ -1,3 +1,8 @@
+- v1.9
+  - Adds support for data parameters.
+  - Adds a new field `triggerRules` to cause spawns from rules to trigger other rules.
+  - Adds a new field `removeDelay` to delay object removing.
+
 - v1.8
   - Adds a new field `drops` to spawn drops.
   - Adds the keyword "creature" to the field `objects` and `bannedObjects`.